style: split task_routes, remove evidence stubs, cargo fmt
Split handle_task_update into task_routes.rs to keep plan_routes.rs
under 250 lines. Removed evidence recording stubs (belongs in
convergio-evidence crate). Applied cargo fmt.
